# Initialialize AABB Rectangle 