--- a/src/BitmapSkin.js
+++ b/src/BitmapSkin.js
@@ -62,10 +62,11 @@ class BitmapSkin extends Skin {
     /**
      * Get the bounds of the drawable for determining its fenced position.
      * @param {Array<number>} drawable - The Drawable instance this skin is using.
+     * @param {?Rectangle} result - Optional destination for bounds calculation.
      * @return {!Rectangle} The drawable's bounds. For compatibility with Scratch 2, we always use getAABB for bitmaps.
      */
-    getFenceBounds (drawable) {
-        return drawable.getAABB();
+    getFenceBounds (drawable, result) {
+        return drawable.getAABB(result);
     }
 
     /**