@mzgoddard 2019-06-14
@adroitwhiz I think #467 and #445 go together. We won't need to cache vectors for getAABB after #467, but we'll still need to cache vectors for getTransformedHullPoints. Probably still use #467 `results` arg and default to the cached `_AABB` instead of creating fresh Rectangles..