@mzgoddard 2019-06-19
I've moved the math proof documentation into a jsdoc tutorial at `docs/Rectangle-AABB-Matrix.md`. It needs some verb tense normalization but it has the proof steps I like.